Initial version.
This is the initial version of the converter, with the following
restrictions:

* Only Spanish numeral conversion, for now (planning to add English and
French numerals as well).
* The accepted numerals ranges from "cero" (0) to "noventa y nueve"
(99), for now.

The next steps are likely to be:

1. Expand the range of accepted Spanish numerals to hundreds, thousands
and millions (things get a little tricky then).
2. Add an English numerals mode, using short scale (en-US, I think), up
to millions.
3. Add an English numerals mode, using long scale (¿en-GB?), up to
millions.
4. Add a French numerals mode, up to millions (this is going to be fun,
considering the way French numerals are built from "soixante-dix" (70)
on.
5. Expand the range of accepted Spanish numerals to weird numbers, like
"millardos" (10^9).
6. Expand the range of accepted numerals in the rest of languages. The
limit is the sky, people!