First version with pattern recursion
This is the first version which recognises numerals from “cero” (0) to
“novecientos noventa y nueve” (999). But, sadly, it won’t work.

The rules `r0` to `r3` (and variants) are all straightforward rules,
with no recursion in them. But the rule `r4` uses in its suffix the
rules `r1` and `r2`… And you cannot use repeated named groups in a
regular expression, at least in Java.

I have two alternatives now, as far as I can see:

* Get rid of the group names, using group numbers instead… But I’m not
even sure it will work.
* Use ANTLR as a grammar parser to recognise Spanish numerals… It seems
to be the way to go, although it implies throwing most of the work done
to the trash bin.

Anyway, the project compiles, but does not work, even for ground rules,
because the compound initial pattern is not compilable.