Major revamp of build system
It’s been several days, but I finally managed to switch the build
system from the Ant/Maven hybrid to a pure Gradle one.

These are the Gradle commands available:

* `gradle`: Compiles the project writes the assembly to the `[dist]`
folder. The assembly comes with two flavours: debug and release.
* `gradle info`: Shows the Maven coordinates of the assembly.
* `gradle publishToMavenLocal`: Publishes the assembly to the local
Maven repo.

Part of the process consists of generating the classes to model the
grammars present in `[src/main/resources/grammars]`.