First working version with ANTLR
It works!

Finally, I have a functional version using ANTLR processing. This
version covers Spanish numerals ranging from “cero” (0) to “novecientos
noventa y nueve” (999).

The main processing class is `SpanishNumeralRecognitionEngine`, which I
have recycled to implement the visitor pattern.

However, the processing is a bit inadequate, for I have no good way to
explore the different alternatives in a compound rule. For example,
`r4`, where the last part could actually fires `r1` or `r2`, not
knowing in advance which one will actually be fired.

With two alternatives it’s not big deal, but when we start processing
huge numbers with compound prefixes and suffixes containing a lot of
rules each, I need a good approach.

But rejoice!