Proper error handling
Almost done!

This is the first complete and functional version with proper error
handling. When you try to convert an invalid Spanish numeral, an
`UnrecognizedNumeralException` is thrown for the caller to capture.

The `README.md` and `CHANGELOG.md` files have been modified
accordingly, with a few minor corrections in the former.